perf(DA): narrow StateRegion propositions to u32 (8 → 4 B/value)
`StateRegion::propositions` carries one `Vec<usize>` per propositional
variable for every footprint and transition-system state. Each inner
element is a concrete value ID bounded by the SAS preprocessor at well
below `u32::MAX`, so storing them as `usize` (8 B) on 64-bit targets is
pure overhead.

Introduce `PropValueId = u32` and switch the inner element type. On SCP
workloads (which keep the per-concrete-op footprints alive) this halves
the propositional bulk content per state region — concretely several
hundred MB on workloads with hundreds of thousands of footprints.

Touch points:
  - `prop_regions_overlap` / `sorted_value_sets_overlap` retyped.
  - `concrete_values_for_abstract_value` / `propositional_region_from_hash`
    / `full_propositional_region` now produce `Vec<u32>` / `Vec<Vec<u32>>`.
  - `singleton_value_for_feature` casts to `usize` at the consumer
    boundary (existing API takes/returns `usize` indices elsewhere).
  - `query_values_for_feature` returns `&[PropValueId]` (callers iterate
    with `&value` and cast at the lone bucket lookup).
  - LM-cut helper `precondition_proposition_ids` reader casts at fact
    construction.

Behaviour unchanged: regression on fillSCP plant-watering and canonical
sailing 1_11 produces the same plan costs as before. Build green, tests
build (pre-existing 7 abstract-operator-generator failures unchanged).

+++ planners-search/src/numeric/evaluation/numeric_landmarks/numeric_lm_cut_landmarks.rs
@@ -859,7 +859,7 @@ impl<'task> LandmarkCutLandmarks<'task> {
         let mut seen = BTreeSet::new();
         for (var, values) in source_region.propositions.iter().enumerate() {
             if values.len() == 1 {
-                let fact = ExplicitFact::new(var, values[0]);
+                let fact = ExplicitFact::new(var, values[0] as usize);
                 for proposition_id in self.precondition_proposition_ids(&fact) {
                     if seen.insert(proposition_id) {
                         ids.push(proposition_id);
